+++ src/PageGalleryPlugin.ts
@@ -3,10 +3,17 @@ import { Notice, Plugin } from 'obsidian'
 import { DataviewApi, getAPI } from 'obsidian-dataview'
 
 import Config from './Config'
+import MemoryTileCache from './MemoryTileCache'
 import PageGalleryRenderChild from './PageGalleryRenderChild'
+import type { TileCache } from './TileWrangler'
 
+// TODO: Figure out if there's a good way to add an on-disk cache so you don't
+// have to fully rebuild the in-memory cache every time the plugin is loaded.
+// Maybe make it optional/configurable, so that users don't have cache files
+// showing up in their workspace if they don't want.
 export default class PageGalleryPlugin extends Plugin {
   api: DataviewApi
+  cache: TileCache
 
   async onload () {
     const api = getAPI()
@@ -16,6 +23,8 @@ export default class PageGalleryPlugin extends Plugin {
       this.showMissingDataviewNotice()
       return
     }
+
+    this.cache = new MemoryTileCache()
 
     this.registerMarkdownCodeBlockProcessor('page-gallery', (source, el, ctx) => this.handlePageGalleryBlock(source, el, ctx))
 
@@ -37,7 +46,13 @@ export default class PageGalleryPlugin extends Plugin {
   async handlePageGalleryBlock (source: string, el: HTMLElement, ctx: MarkdownPostProcessorContext) {
     try {
       const config = Config.parse(source)
-      const child = new PageGalleryRenderChild(this, config, this.api, el)
+      const child = new PageGalleryRenderChild({
+        plugin: this,
+        config,
+        api: this.api,
+        element: el,
+        cache: this.cache
+      })
       ctx.addChild(child)
     } catch (err) {
       const pre = document.createElement('pre')